# controller/DepartamentoController.go
package controller

import (
	"context"
	"errors"
	"net/http"
	"strconv"

	"github.com/davi-fernandesx/sistema-de-gestao-de-epi/internal/helper"
	"github.com/davi-fernandesx/sistema-de-gestao-de-epi/internal/model"
	"github.com/gin-gonic/gin"
)

type DepartamentoService interface {
	SalvarDepartamento(ctx context.Context, model model.Departamento) error
	ListarDepartamento(ctx context.Context, id int32) (model.DepartamentoDto, error)
	ListarTodosDepartamentos(ctx context.Context) ([]model.DepartamentoDto, error)
	DeletarDepartamento(ctx context.Context, id int) error
	AtualizarDepartamento(ctx context.Context, id int32, novoNome string) error
}

type DepartamentoController struct {
	service DepartamentoService
}

func NewDepartamentoController(service DepartamentoService) *DepartamentoController {

	return &DepartamentoController{service: service}
}

// RegistraDepartamento godoc
// @Summary      Criar novo departamento
// @Description  Cadastra um novo departamento no sistema
// @Tags         Departamentos
// @Accept       json
// @Produce      json
// @Param        departamento body model.Departamento true "Dados do departamento"
// @Success      201  {object}  map[string]string
// @Failure      400  {object}  helper.HTTPError "Dados inválidos"
// @Failure      409  {object}  helper.HTTPError "Departamento já existe"
// @Failure      500  {object}  helper.HTTPError "Erro interno"
// @Router       /departamentos [post]
// @Security     BearerAuth
func (d *DepartamentoController) RegistraDepartamento() gin.HandlerFunc {

	return func(c *gin.Context) {

		var input model.Departamento

		if err := c.ShouldBindJSON(&input); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{
				"error":    "dados invalidos",
				"detalhes": err.Error(),
			})
			return
		}

		novoDep := model.Departamento{
			Departamento: input.Departamento,
		}

		err := d.service.SalvarDepartamento(c, novoDep)
		if err != nil {

			if errors.Is(err, helper.ErrDadoDuplicado) {
				c.JSON(http.StatusConflict, gin.H{

					"error": err.Error(),
				})
				return
			}
			c.JSON(http.StatusInternalServerError, gin.H{

				"error": err.Error(),
			})
			return
		}

		c.JSON(http.StatusCreated, gin.H{

			"mensagem": "departamento cadastrado",
		})
	}
}

// ListarDepartamentos godoc
// @Summary      Listar todos
// @Description  Retorna uma lista com todos os departamentos
// @Tags         Departamentos
// @Produce      json
// @Success      200  {array}   model.DepartamentoDto
// @Failure      500  {object}  helper.HTTPError "Erro interno"
// @Router       /departamentos [get]
// @Security     BearerAuth
func (d *DepartamentoController) ListarDepartamentos() gin.HandlerFunc {

	return func(ctx *gin.Context) {

		deps, err := d.service.ListarTodosDepartamentos(ctx.Request.Context())
		if err != nil {

			ctx.JSON(http.StatusInternalServerError, gin.H{
				"error": "Erro interno ao listar departamentos",
			})
			return
		}

		ctx.JSON(http.StatusOK, deps)
	}
}

// ListarDepartamentoId godoc
// @Summary      Buscar por ID
// @Description  Retorna os detalhes de um único departamento
// @Tags         Departamentos
// @Produce      json
// @Param        id   path      int  true  "ID do Departamento"
// @Success      200  {object}  model.DepartamentoDto
// @Failure      400  {object}  helper.HTTPError "ID inválido"
// @Failure      404  {object}  helper.HTTPError "Não encontrado"
// @Failure      500  {object}  helper.HTTPError "Erro interno"
// @Router       /departamentos/{id} [get]
// @Security     BearerAuth
func (d *DepartamentoController) ListarDepartamentoId() gin.HandlerFunc {

	return func(ctx *gin.Context) {

		idString := ctx.Param("id")

		id, err := strconv.Atoi(idString)
		if err != nil {

			ctx.JSON(http.StatusBadRequest, gin.H{
				"error": "id deve ser um numero",
			})
			return
		}

		dep, err := d.service.ListarDepartamento(ctx, int32(id))
		if err != nil {

			if errors.Is(err, helper.ErrNaoEncontrado) {

				ctx.JSON(http.StatusNotFound, gin.H{

					"error": "departamento nao encontrado",
				})
				return
			}

			ctx.JSON(http.StatusInternalServerError, gin.H{

				"error": err.Error(),
			})
			return

		}

		ctx.JSON(http.StatusOK, dep)

	}
}

// DeletarDepartamento godoc
// @Summary      Deletar departamento
// @Description  Remove (ou inativa) um departamento pelo ID
// @Tags         Departamentos
// @Param        id   path      int  true  "ID do Departamento"
// @Success      204  "Sem Conteúdo (Sucesso)"
// @Failure      400  {object}  helper.HTTPError "ID inválido"
// @Failure      404  {object}  helper.HTTPError "Não encontrado"
// @Failure      500  {object}  helper.HTTPError "Erro interno"
// @Router       /departamentos/{id} [delete]
// @Security     BearerAuth
func (d *DepartamentoController) DeletarDepartamento() gin.HandlerFunc {

	return func(ctx *gin.Context) {

		idString := ctx.Param("id")

		id, err := strconv.Atoi(idString)
		if err != nil {

			ctx.JSON(http.StatusBadRequest, gin.H{
				"error": "id deve ser um numero",
			})
			return
		}

		err = d.service.DeletarDepartamento(ctx, id)
		if err != nil {

			if errors.Is(err, helper.ErrNaoEncontrado) {

				ctx.JSON(http.StatusNotFound, gin.H{

					"error": "departamento nao encontrado",
				})

				return
			}

			ctx.JSON(http.StatusInternalServerError, gin.H{

				"error": err.Error(),
			})
			return
		}

		ctx.Status(http.StatusNoContent)
	}
}

// UpdateDepartamento godoc
// @Summary      Atualizar departamento
// @Description  Atualiza o nome de um departamento existente
// @Tags         Departamentos
// @Accept       json
// @Produce      json
// @Param        id   path      int                      true  "ID do Departamento"
// @Param        body body      model.Departamento true  "Novo nome"
// @Success      200  {object}  map[string]string "Sucesso"
// @Failure      400  {object}  helper.HTTPError "Erro de validação (ID ou Nome curto)"
// @Failure      404  {object}  helper.HTTPError "Não encontrado"
// @Failure      500  {object}  helper.HTTPError "Erro interno"
// @Router       /departamentos/{id} [put]
func (d *DepartamentoController) AtualizarDepartamento() gin.HandlerFunc {

	return func(ctx *gin.Context) {

		idParam := ctx.Param("id")
		id, err := strconv.Atoi(idParam)
		if err != nil {

			ctx.JSON(http.StatusBadRequest, gin.H{
				"erro": err.Error(),
			})
			return
		}

		var input model.Departamento

		if err := ctx.ShouldBindJSON(&input); err != nil {

			ctx.JSON(http.StatusBadRequest, gin.H{

				"error": err.Error(),
			})
			return
		}

		err = d.service.AtualizarDepartamento(ctx, int32(id), input.Departamento)
		if err != nil {

			if errors.Is(err, helper.ErrNomeCurto) {

				ctx.JSON(http.StatusNotFound, gin.H{

					"error": "nome do departamento tem que possui 2 ou mais letras",
				})
				return
			}
			if errors.Is(err, helper.ErrNaoEncontrado) {

				ctx.JSON(http.StatusNotFound, gin.H{

					"error": "departamento nao encontrado para atualizar",
				})
				return
			}

			ctx.JSON(http.StatusInternalServerError, gin.H{

				"erro": err.Error(),
			})

			return
		}

		ctx.JSON(http.StatusOK, gin.H{"sucesso": "departamento atualizado"})
	}
}