# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestão de EPIs e Departamentos.",
        "title": "SaaS EPI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.seusaas.com.br",
            "email": "suporte@seusaas.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/departamentos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna uma lista com todos os departamentos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Listar todos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DepartamentoDto"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cadastra um novo departamento no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Criar novo departamento",
                "parameters": [
                    {
                        "description": "Dados do departamento",
                        "name": "departamento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Departamento"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Departamento já existe",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/departamentos/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os detalhes de um único departamento",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Buscar por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Departamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DepartamentoDto"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza o nome de um departamento existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departamentos"
                ],
                "summary": "Atualizar departamento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Departamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novo nome",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Departamento"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação (ID ou Nome curto)",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove (ou inativa) um departamento pelo ID",
                "tags": [
                    "Departamentos"
                ],
                "summary": "Deletar departamento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Departamento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sem Conteúdo (Sucesso)"
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helper.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Departamento": {
            "type": "object",
            "required": [
                "departamento"
            ],
            "properties": {
                "departamento": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "model.DepartamentoDto": {
            "type": "object",
            "properties": {
                "departamento": {
                    "type": "string",
                    "example": "Recursos Humanos"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}